+++ packages/renderer/src/components/ButtonColumn.vue
@@ -1,9 +1,9 @@
 <script setup lang="ts">
 import { inject, ref, computed } from "vue";
 import type { Column } from "interfaces";
-import type Store from "../store/Store";
+import type { Store } from "../store";
 import BaseButtonToggle from "./BaseButtonToggle.vue";
-import useIsSidebarDisabled from "../composables/useIsSidebarDisabled";
+import useIsSidebarDisabled from "../composables/use-is-sidebar-disabled";
 
 const store = inject("store") as Store;
