+++ packages/renderer/src/composables/use-is-sidebar-disabled.ts
