+++ packages/renderer/src/store/index.ts
@@ -1,9 +1,18 @@
-import Store from "./Store";
-import ProjectStore from "./modules/ProjectStore";
-import ApplicationStore from "./modules/ApplicationStore";
+import ProjectStore from "./store-project";
+import ApplicationStore from "./store-application";
 
 const applicationStore = new ApplicationStore();
 const projectStore = new ProjectStore();
+
+export class Store {
+  public application: ApplicationStore;
+  public projects: ProjectStore;
+
+  constructor(application: ApplicationStore, projectStore: ProjectStore) {
+    this.application = application;
+    this.projects = projectStore;
+  }
+}
 
 const store = new Store(applicationStore, projectStore);
