+++ packages/renderer/src/store/store-application.ts
