+++ packages/renderer/src/views/Welcome.vue
@@ -5,7 +5,7 @@ import BaseCardContent from "../components/BaseCardContent.vue";
 import BaseButtonClick from "../components/BaseButtonClick.vue";
 import FormProjectCreateModal from "../components/FormProjectCreateModal.vue";
 import NotificationDot from "../components/NotificationDot.vue";
-import type Store from "../store/Store";
+import type { Store } from "../store";
 
 const store = inject("store") as Store;
 const isModalActive = ref<boolean>(false);
